Fix verbose option argument parsing (#8, #11)
Fix #8: macrodir: verbose is on by default, option -v|--vervose is no-op
Fix #11: macrodir: Option -v eats diretory name if it looks like a number

Verbose option was being initialized with 1 (i.e. verbose on) and accepted an
optional number to set verbose to. If the directory was a number, that number
was used as the verbose level.

Verbose level two was used to show files up to date.

Changed verbose to be a flag without arguments (0/1), initialized to 0.