Use Object::Tiny::RW to create object and access attributes
Rewrite new() to create the object based on Object::Tiny::RW, $self->atrtribute to read
and $self->attribute($new_value) to write; still use the old $self->{attribute} when
required, e.g. inside regexps, as lvalue, and to access MACRO, VARIABLE or SCRIPT
depending on a $which variable.

Do not store the passed in %options in the object during new(), as the options are only used
to initialize the object.